Rename rattt variables and directories to ralambda (#37)
* rename package rattt to ralambda

* rename rattt to ralambda for tests

* replace additional occurrences of ttt with lambda
