+++ src/main/java/de/learnlib/ralib/learning/ralambda/RaDT.java
@@ -1,4 +1,4 @@
-package de.learnlib.ralib.learning.rattt;
+package de.learnlib.ralib.learning.ralambda;
 
 import java.util.Deque;
 import java.util.LinkedHashMap;
