+++ src/test/java/de/learnlib/ralib/learning/ralambda/LearnEchoTest.java
@@ -1,7 +1,8 @@
-package de.learnlib.ralib.learning.rattt;
+package de.learnlib.ralib.learning.ralambda;
 
 import static de.learnlib.ralib.example.repeater.RepeaterSUL.IPUT;
 import static de.learnlib.ralib.example.repeater.RepeaterSUL.OECHO;
+import static de.learnlib.ralib.example.repeater.RepeaterSUL.ONOK;
 import static de.learnlib.ralib.example.repeater.RepeaterSUL.TINT;
 
 import java.util.LinkedHashMap;
@@ -18,8 +19,6 @@ import de.learnlib.ralib.data.DataType;
 import de.learnlib.ralib.data.DataValue;
 import de.learnlib.ralib.example.repeater.RepeaterSUL;
 import de.learnlib.ralib.learning.Hypothesis;
-import de.learnlib.ralib.learning.Measurements;
-import de.learnlib.ralib.learning.QueryStatistics;
 import de.learnlib.ralib.oracles.SimulatorOracle;
 import de.learnlib.ralib.oracles.TreeOracleFactory;
 import de.learnlib.ralib.oracles.io.IOCache;
@@ -35,10 +34,10 @@ import de.learnlib.ralib.tools.theories.IntegerEqualityTheory;
 import de.learnlib.ralib.words.PSymbolInstance;
 import net.automatalib.words.Word;
 
-public class TestSuffixOptimization extends RaLibTestSuite {
+public class LearnEchoTest extends RaLibTestSuite {
 
-	@Test
-	public void learnRepeater() {
+    @Test
+    public void learnEchoTest() {
 
         Constants consts = new Constants();
 
@@ -47,45 +46,40 @@ public class TestSuffixOptimization extends RaLibTestSuite {
         theory.setUseSuffixOpt(true);
         teachers.put(TINT, theory);
 
-        RepeaterSUL sul = new RepeaterSUL(-1, 2);
+        RepeaterSUL sul = new RepeaterSUL(-1, 4);
         IOOracle ioOracle = new SULOracle(sul, RepeaterSUL.ERROR);
-	    IOCache ioCache = new IOCache(ioOracle);
-	    IOFilter oracle = new IOFilter(ioCache, sul.getInputSymbols());
+	IOCache ioCache = new IOCache(ioOracle);
+	IOFilter oracle = new IOFilter(ioCache, sul.getInputSymbols());
 
         ConstraintSolver solver = new SimpleConstraintSolver();
 
-        MultiTheoryTreeOracle mto = new MultiTheoryTreeOracle(
-                oracle, teachers, consts, solver);
-        MultiTheorySDTLogicOracle mlo =
-                new MultiTheorySDTLogicOracle(consts, solver);
+        MultiTheoryTreeOracle mto = new MultiTheoryTreeOracle(oracle, teachers, consts, solver);
+        MultiTheorySDTLogicOracle mlo = new MultiTheorySDTLogicOracle(consts, solver);
 
         TreeOracleFactory hypFactory = (RegisterAutomaton hyp) ->
                 new MultiTheoryTreeOracle(new SimulatorOracle(hyp), teachers, consts, solver);
 
-        Measurements measurements = new Measurements();
-        QueryStatistics queryStats = new QueryStatistics(measurements, sul);
-
         RaLambda learner = new RaLambda(mto, hypFactory, mlo, consts, true, sul.getActionSymbols());
-        learner.setStatisticCounter(queryStats);
-        learner.setSolver(solver);
-
         learner.learn();
 
         Word<PSymbolInstance> ce = Word.fromSymbols(
-        		new PSymbolInstance(IPUT, new DataValue(TINT, 0)),
-        		new PSymbolInstance(OECHO, new DataValue(TINT, 0)),
-        		new PSymbolInstance(IPUT, new DataValue(TINT, 1)),
-        		new PSymbolInstance(OECHO, new DataValue(TINT, 1)),
-        		new PSymbolInstance(IPUT, new DataValue(TINT, 2)),
-        		new PSymbolInstance(OECHO, new DataValue(TINT, 2)));
+       	        new PSymbolInstance(IPUT, new DataValue(TINT, 0)),
+       	        new PSymbolInstance(OECHO, new DataValue(TINT, 0)),
+       	        new PSymbolInstance(IPUT, new DataValue(TINT, 1)),
+       	        new PSymbolInstance(OECHO, new DataValue(TINT, 1)),
+       	        new PSymbolInstance(IPUT, new DataValue(TINT, 2)),
+       	        new PSymbolInstance(OECHO, new DataValue(TINT, 2)),
+       	        new PSymbolInstance(IPUT, new DataValue(TINT, 3)),
+       	        new PSymbolInstance(OECHO, new DataValue(TINT, 3)),
+       	        new PSymbolInstance(IPUT, new DataValue(TINT, 4)),
+       	        new PSymbolInstance(ONOK));
 
-        learner.addCounterexample(new DefaultQuery<PSymbolInstance, Boolean>(ce, false));
-
+        learner.addCounterexample(new DefaultQuery<>(ce, true));
         learner.learn();
+
         Hypothesis hyp = learner.getHypothesis();
 
-        Assert.assertEquals(hyp.getStates().size(), 7);
-
-        System.out.println(queryStats.toString());
-	}
+        Assert.assertEquals(hyp.getStates().size(), 11);
+        Assert.assertTrue(hyp.accepts(ce));
+    }
 }
