+++ src/test/java/de/learnlib/ralib/learning/ralambda/LearnLoginTest.java
@@ -1,4 +1,4 @@
-package de.learnlib.ralib.learning.rattt;
+package de.learnlib.ralib.learning.ralambda;
 
 import static de.learnlib.ralib.example.login.LoginAutomatonExample.AUTOMATON;
 import static de.learnlib.ralib.example.login.LoginAutomatonExample.I_LOGIN;
@@ -62,12 +62,12 @@ public class LearnLoginTest extends RaLibTestSuite {
                 new MultiTheoryTreeOracle(new SimulatorOracle(hyp), teachers,
                         new Constants(), solver);
 
-        RaLambda rattt = new RaLambda(mto, hypFactory, slo,
+        RaLambda ralambda = new RaLambda(mto, hypFactory, slo,
                 consts, I_LOGIN, I_LOGOUT, I_REGISTER);
-        rattt.setSolver(solver);
+        ralambda.setSolver(solver);
 
-        rattt.learn();
-        RegisterAutomaton hyp = rattt.getHypothesis();
+        ralambda.learn();
+        RegisterAutomaton hyp = ralambda.getHypothesis();
         logger.log(Level.FINE, "HYP1: {0}", hyp);
 
         Word<PSymbolInstance> ce = Word.fromSymbols(
@@ -76,10 +76,10 @@ public class LearnLoginTest extends RaLibTestSuite {
                 new PSymbolInstance(I_LOGIN,
                         new DataValue(T_UID, 0), new DataValue(T_PWD, 0)));
 
-        rattt.addCounterexample(new DefaultQuery<>(ce, sul.accepts(ce)));
+        ralambda.addCounterexample(new DefaultQuery<>(ce, sul.accepts(ce)));
 
-        rattt.learn();
-        hyp = rattt.getHypothesis();
+        ralambda.learn();
+        hyp = ralambda.getHypothesis();
         logger.log(Level.FINE, "HYP2: {0}", hyp);
 
         Assert.assertEquals(hyp.getStates().size(), 3);
@@ -100,7 +100,7 @@ public class LearnLoginTest extends RaLibTestSuite {
         teachers.put(T_UID, new IntegerEqualityTheory(T_UID));
         teachers.put(T_PWD, new IntegerEqualityTheory(T_PWD));
 
-        Measurements[] measuresTTT = new Measurements[SEEDS];
+        Measurements[] measuresLambda = new Measurements[SEEDS];
         Measurements[] measuresStar = new Measurements[SEEDS];
 
         RaLibLearningExperimentRunner runner = new RaLibLearningExperimentRunner(logger);
@@ -108,7 +108,7 @@ public class LearnLoginTest extends RaLibTestSuite {
         for (int seed=0; seed<SEEDS; seed++) {
         	runner.setSeed(seed);
 	        Hypothesis hyp = runner.run(RaLearningAlgorithmName.RALAMBDA, dwOracle, teachers, consts, solver, new ParameterizedSymbol [] {I_LOGIN, I_LOGOUT, I_REGISTER});
-	        measuresTTT[seed] = runner.getMeasurements();
+	        measuresLambda[seed] = runner.getMeasurements();
 	        runner.resetMeasurements();
 
 	        Assert.assertEquals(hyp.getStates().size(), 4);
@@ -119,7 +119,7 @@ public class LearnLoginTest extends RaLibTestSuite {
 	        measuresStar[seed] = runner.getMeasurements();
 	        runner.resetMeasurements();
         }
-        System.out.println("Queries (TTT): " + Arrays.toString(measuresTTT));
+        System.out.println("Queries (Lambda): " + Arrays.toString(measuresLambda));
         System.out.println("Queries (Star): " + Arrays.toString(measuresStar));
     }
